refactor: 훅과 컴포넌트들을 분리.
